+++ src/main/java/pl/zdusza/Device.java
@@ -1,0 +1,5 @@
+package pl.zdusza;
+
+public enum Device {
+    IPHONE, GENERIC_MOBILE, RICH_CSS, TABLET, OTHER
+}