+++ src/main/java/pl/zdusza/event/EventLogMarker.java
@@ -1,0 +1,5 @@
+package pl.zdusza.event;
+
+public enum EventLogMarker {
+    MONITORING_EVENT,
+}