+++ src/test/java/pl/zdusza/DeviceTest.java
@@ -1,0 +1,11 @@
+package pl.zdusza;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class DeviceTest {
+    @Test
+    public final void testShouldEnumValuesEqualsEnumValues() {
+        Assert.assertArrayEquals(Device.values(), Device.values());
+    }
+}