+++ src/test/java/pl/zdusza/event/EventLogMarkerTest.java
@@ -1,0 +1,11 @@
+package pl.zdusza.event;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class EventLogMarkerTest {
+    @Test
+    public final void testShouldEnumValuesEqualsEnumValues() {
+        Assert.assertArrayEquals(EventLogMarker.values(), EventLogMarker.values());
+    }
+}