+++ src/test/java/pl/zdusza/time/TimerTest.java
@@ -6,7 +6,7 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
-import org.slf4j.Logger;
+import pl.zdusza.event.EventLogger;
 import pl.zdusza.event.LoggableEventName;
 
 import java.time.Duration;
@@ -24,7 +24,7 @@ public class TimerTest {
     public MockitoRule mockitoRule = MockitoJUnit.rule();
 
     @Mock
-    private Logger logger;
+    private EventLogger logger;
 
     private final Clock clock = new Clock(Instant.parse("2019-01-01T00:31:00.000Z"));
 
@@ -35,11 +35,9 @@ public class TimerTest {
         clock.setNow(Instant.parse("2019-01-01T00:31:00.001Z"));
         timer.stop();
         Mockito.verify(logger, Mockito.times(1))
-                .info("{} ms:{} roundeds:{} context:{}",
-                        LoggableEventName.RESPONSE_TIME,
-                        RESPONSE_TIME_IN_MILLIS,
-                        RESPONSE_TIME_ROUNDED_TO_SEC,
-                        TimerTest.class.getName());
+                .info(LoggableEventName.RESPONSE_TIME, TimerTest.class.getName(),
+                        "ms:{} roundeds:{}", RESPONSE_TIME_IN_MILLIS,
+                        RESPONSE_TIME_ROUNDED_TO_SEC);
     }
 
     @Test
@@ -49,10 +47,8 @@ public class TimerTest {
         clock.setNow(Instant.parse("2019-01-01T00:31:01.001Z"));
         timer.stop();
         Mockito.verify(logger, Mockito.times(1))
-                .error("{} ms:{} roundeds:{} context:{}",
-                        LoggableEventName.HIGH_RESPONSE_TIME,
-                        HIGH_RESPONSE_TIME_IN_MILLIS,
-                        HIGH_RESPONSE_TIME_ROUNDED_TO_SEC,
-                        TimerTest.class.getName());
+                .error(LoggableEventName.HIGH_RESPONSE_TIME, TimerTest.class.getName(),
+                        "ms:{} roundeds:{}", HIGH_RESPONSE_TIME_IN_MILLIS,
+                        HIGH_RESPONSE_TIME_ROUNDED_TO_SEC);
     }
 }