+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] - 2019-04-04
+### Added
+- EventLogger, EventLogMarker, Period, StorableEventName
+### Changed
+- Timer interface to get EventLogger instead of Logger
+
 ## [2.1.0] - 2019-03-14
 ### Added
 - Role again